Allow one of our projects to be missing the totalBilledBytes field
Summary:
The gae_dashboard_stats job has been failing on toby, because the
bigquery logs export table for the khan-academy project is missing the
...totalBilledBytes field.  Looking at the list of jobs for that
project, it seems like we just haven't had a successful job run in that
project yet today, so we don't actually have that field in the logs yet.

This revision makes it so that we ignore (logging a warning) that error
if it happens for some subset (i.e. not all) of our projects being cost
monitored.  My rationale here was that it's much more unlikely we
haven't had a successful job today in any of our projects, so if we see
this condition, it may mean something is wrong that needs human
intervention.

Test Plan:
- `gae_dashboard/fetch_bigquery_cost.py --dry-run --date 20170614`
- verify that since it is currently missing the field only in the
  khan-academy project, it logs a warning but continues, and that it
  says it would send the stats to graphite.
- remove the deductive-jet project from the list of projects to check,
  so that khan-academy is now the only project
- repeat the command, verify it now raises an exception because all
  projects failed with this message

Reviewers: nabil

Reviewed By: nabil

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D36556