Use `GROUP EACH BY` for the OOM query in email_bq_data.py
Summary:

See https://app.asana.com/0/31965416896056/366173772060621

This commit fixes an issue where the query to calculate OOMs by route is
failing with "resources exceeded".

The query in question doesn't do any joins, which are the usual
culprit, so the only remaining candidates are the `GROUP BY` and
`ORDER BY` clauses.  Neither changing the `GROUP BY` to `GROUP EACH BY`
nor removing the `ORDER BY` helped, so the only remaining thing to do
is to decrease the amount of data we keep in memory relative to the
resources allocated.

The most straightforward way to do this was to remote the
`NEST(app_logs.message) WITHIN RECORD` that I added in D35735, and
replace it with a less general solution for each query that works with
something that's smaller than the full complement of messages for each
record.

For the OOM query, I replaced it with summing integers for whether each
message contains the OOM string.  For the memory profiling query (which
also hit resources exceeded after I fixed the first one), I replaced it
with summing up the memory total and added figures across all messages
in the record (which should all be 0, except for the message where the
memory profiling info appears).

Test Plan:
- `./email_bq_data.py --dry-run --date 20170625`
- verify this completes successfully, whereas before it would hit
  resources exceeded every time.

Auditors: benkraft, csilvers