Add react_render_server.failure_pct as a stat we wsend over.
We have to go through graphite for this because stackdriver can't do
the ratios manually.

If I was better at graphite I'd be able to send over failure
information for the most failing-est routes as well.  We'd need to add
metric-labels to graphite_bridge, though.

Auditors: nabil, colin

Test plan:
Fingers crossed.  I'll try to run this manually to at least make sure
it doesn't crash.