Use the "schema_update_option" to allow adding new sailthru campaign fields
Summary:

The sailthru export to bigquery is failing because it can't change the
schema on the table it's trying to write (the campaigns table).  I
think that if we specify the
`--schema_update_option=ALLOW_FIELD_ADDITION` flag it will
automatically patch the schema as new fields get added.

Test Plan:
- fingers crossed (we're about to have a daily run of this, so we can
just try it out on that)

Auditors: ragini