Make email reliability metrics report more robust against missing data.
Summary:
Prior to this commit, email_reliabilty_metrics.py implicitly required that all metrics
were present in every week of data.  This was not true this week as `signup_page` data was missing
due to changes on the frontend, resulting in the script failing.
This commit changes the graphing function so that it will graph whatever data is present
rather than break if a week of data is missing.
This also means that we can add new metrics without backfilling data or discarding old data.

I also rotated the x-axis labels vertically so that the dates are actually readable.

Test Plan:
I downloaded the email_reliability_metrics.db from toby via sftp,
ran the script with the `--dry-run` and `--save-graphs` options,
and observed that the output graphs looked as I expected.

Reviewers: amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D38379