Fix logs bridge
Summary:
Currently logs bridge isn't sending data to stackdriver.

https://app.asana.com/0/31965416896056/391517463351604

The problem is that some bq tables have moved. This diff
changes which tables the logs bridge script queries.

I tried both querying both the last_hour table and the
hourly tables. They give similar results for recent times.

It seems like there is currently a small delay of a few minutes
before logs arrive in the hourly tables, based on looking at
time stamps of the hourly logs in bq. I decided to use 45 minutes
as cut off of when we expect logs to have arrived in the
hourly table, which hopefully is enough in most circumstances.

Test Plan:
I ran::

    python logs_bridge.py -v --dry-run

Noted that it worked with these changes but fails without them. Noted
that the failure was due to referencing a non-existent table.

Reviewers: nabil

Reviewed By: nabil

Differential Revision: https://phabricator.khanacademy.org/D38980