Add a new daily email that gives stats on rrs timeouts and errors.
Summary:
Investigating react render server problems we've noticed that there
are lots of components that take too long to render. For components
that aren't able to render in less than a second, the render is timed
out and we end up adding a second to the request and rendering on the
client.

We've also noticed that some components frequently fail to render.

This report should help the relevant initiatives see which components
are having problems. Fixing these problems should give decent page speed-ups.

Test Plan:
- Manually tested locally by running:
  python gae_dashboard/email_bq_data.py --dry-run --report=email_rrs_stats

Reviewers: drosile, nabil

Reviewed By: nabil

Differential Revision: https://phabricator.khanacademy.org/D39937