Change emails that some reports are sent to.
Summary:
We are no longer going to send out the weekly perf reports to the
whole company. For now send it only to infra blackhole.

Also content platform wants their daily reports to go to a separate
email list. I created content-platform-analytics for that purpose.

Test Plan: Hope for the best.

Reviewers: nabil

Reviewed By: nabil

Differential Revision: https://phabricator.khanacademy.org/D40169