Improve rrs latency reports
Summary:
The current latency reports don't give a clear picture
of timeouts.

Now we count timeouts and show the percentage of reqs
that timeout.

Also removed best latency from latency report, since it's
not very useful now that we have timeout percentage.

Also changed error report to not report number of ok
requests, since this will be a bit misleading with the
new way that rrs will use 504s for timeouts.

Test Plan:
Test locally with:
    python gae_dashboard/email_bq_data.py --dry-run --report=email_rrs_stats

Reviewers: nabil

Reviewed By: nabil

Differential Revision: https://phabricator.khanacademy.org/D40267