Check for failing routes.
Summary:
https://khanacademy.atlassian.net/browse/INFRA-101
https://app.asana.com/0/31965416896056/312222409070988

Sometimes routes stop giving 2xx responses. If the
routes aren't hit very often it can be hard to notice.

The diff adds a script that should be called by cron
once a day to look for routes that haven't given
any 2xx responses in the last day. Note: we allow
allow 3xx and 401, 404, 405 since they don't
indicated that the route is broken.

If routes are found that have had responses but not
ok responses then alerts are sent to #infrastructure-alerts
and to alerta.

This diff also updates the alertlib-src submodule to
origin master so that it picks up the ability to
send to alerta. I don't think that this is used
in prod, but it helps for local testing.

Deploy Plan:

Updated internal-webserver on toby.

Create a new entry in aws-config toby/crontab to call
the script once a day.

Test Plan:
I manually tested the case where there are no failing
routes by changing the query to return no failing routes:

    python gae_dashboard/check_failing_routes.py --dry-run

The output was:

    No routes with no 2xx requests for 10/30/17

Also ran it without dry-run:

    python gae_dashboard/check_failing_routes.py

I noticed that no alerts were sent to alerta or slack.

Then I tested with dry-run:

    python gae_dashboard/check_failing_routes.py --dry-run

The output was:

    Routes with no 2xx requests for 10/31/17:
    main:/<subject_slug:.*>/<slug>/<scratchpad_id:\d+>/embedded/<trusted:trusted>
    api.main:/api/internal/badges/award [POST]
    main:/stats/memory

I tested without dry-run:

    python gae_dashboard/check_failing_routes.py

I noticed that alerts were sent to alerta and slack. Slack had a single
alert and there were three alerts in alerta. Each alert in alerta
had the correct initiative.

Reviewers: nabil, jacqueline, drosile

Reviewed By: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D40556