Use our fork of sailthru-api-client, that supports timeout and retries.
Summary:
We're hoping these get accepted upstream:
   https://github.com/sailthru/sailthru-python-client/pull/67
but no need to wait...

I also got rid of some requirements that aren't used anymore, which I
figured by doing
    git grep -e bigquery -e cssselect -e ez-setup -e lxml
and seeing no results.

Resolves https://app.asana.com/0/31965416896056/453350291157031

Test Plan:
I ran
   pip install -r requirements.txt
and then
   cat /home/csilvers/.virtualenv/khan27/lib/python2.7/site-packages/sailthru/sailthru_http.py
and saw the timeout and retries arg to sailthru_http_request().

But the real test will be to install this on toby (which I'll do by
logging in, doing a git pull in internal-webserver/, and then manually
running the pip install).

Reviewers: ragini

Reviewed By: ragini

Differential Revision: https://phabricator.khanacademy.org/D40599