Fix false positives in failing routes report
Summary:
Today we saw a couple false positives in the
failing routes report:

    main:<none> [PROPFIND]
    main:<none> [LOCK]

It turns out that you get a 501 (not a 405) when
requesting these methods. So I've updated the list of
OK status codes to include 501, since it doesn't
indicate an error on the part of the route, but
rather on the part of the client.

Test Plan:
Locally I ran:

    python gae_dashboard/check_failing_routes.py --dry-run

without the change and saw the false positives:

    main:<none> [PROPFIND]
    api.main:/api/internal/videos/<youtube_id>/localized/<ka_locale>/<translated_id> [PUT]
    api.main:/api/internal/dev/publish_content/<int:status_id>
    /api/internal/dev/edit/LearnMenuCuration/<content_id> [PUT]
    main:<none> [LOCK]

After the changes from this diff the same command gives the
same output only without the false positives:

    api.main:/api/internal/videos/<youtube_id>/localized/<ka_locale>/<translated_id> [PUT]
    /api/internal/dev/edit/LearnMenuCuration/<content_id> [PUT]
    api.main:/api/internal/dev/publish_content/<int:status_id>

Reviewers: nabil, jacqueline, drosile

Reviewed By: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D40688