Only consider user-visible modules for "degraded performance"
Summary:
Sadly, we have a lot of 5xx errors on react-render.  These aren't great
but we shouldn't be counting them as "degraded performance" in the
reliability reports.  More generally errors on non-user-facing modules
don't necessarily indicate a problem users can see.  (Better might be to
distinguish which batch jobs users can quickly see the results of, but
we don't have that documented.)  So we now only count the modules
end-users talk to directly.

I also updated the top-of-file docstring bit for how the script is
currently used, while I was in the area.

Test Plan:
I ran the fixed query for 2017-11-08, and saw about 3 minutes of
downtime reported instead of about 6 hours.

Reviewers: amos, csilvers

Reviewed By: amos, csilvers

Subscribers: kevindangoor

Differential Revision: https://phabricator.khanacademy.org/D40820