Disable sending RRS stats email when no logs
Summary:
Now that RRS had been trned off, we do not get any new tables (of type
react_render_logs.appengine_googleapis_com_nginx_request_{}) generated
for the RRS logs. Look at
https://khanacademy.atlassian.net/browse/INFRA-400 for more details.

The solution adopted is: Check if the RRS logs table exists. If not,
skip the rest of the logic to send an email about the RRS stats. This
will work as is when RRS is enabled (if..). This takes into account that
the email is not sent only in the case when there is no corresponding
RRS logs table. Ideally, we could extend this logic to other tables but
apparently we do not need to as all other stats query request logs which
are always there (hopefully).

Test Plan: Ran the email_rrs_stats script for the dates when the table is absent as well as when it is present in the dry-run mode to verify it does as expected.

Reviewers: benkraft, amos

Differential Revision: https://phabricator.khanacademy.org/D40950