Ignore repo uris that aren't observing anything.
Joshua N noticed that github is case-insensitive but phabricator is
case-sensitive.  So if someone cloned github.com:khan/webapp, phab
wouldn't know what that was (since it only knows about Khan/webapp).
So he added an "alias" uri to the phab config to let it know about
both forms.

This changes the script to ignore that alias uri when deciding what
the list of phabricator repos is.  It now only pays attention to uris
that are "observing" a github repo (which respects the canonical
capitalization that github has).  In theory we could do better to
ensure that we only paid attention to one uri per repo, but this
should be good enough in practice.

It's still a TODO to automatically set up the alias uri when adding a
new repo.  (We may also want to automatically set up a staging area
for each new repo as well, which is what motivated joshua's
investigation in the first place.)  But this will quiet the 5-minutely
cron job at least.

Auditors: jared, joshua

Test plan:
   % ./update_phabricator_repositories.py -n /tmp
   Adding 0 new repositories to phabricator
   Removing 0 deleted repositories from phabricator