Change location of tempdir for sailthru exports
Summary:
The temporary files that are created during the Sailthru to BigQuery
exports occupy too much space. Currently they are created in the /tmp
directory. We now change the path of these temporary files to be in the
/mnt/sailthru_tmp directory. I manually created a directory called
sailthru_tmp in /mnt on toby and assigned permissions to it to be able
to create temporary files in it.

Test Plan: fingers crossed

Reviewers: csilvers

Subscribers: #infrastructure