Updates and clarifications to the instance hour report
Summary:
I got rather confused at discrepancies between the number of instance
hours stackdriver was reporting and our instance hours reports. It
turns out this is because we're calculating instance hours normalized to
the cost of a perfect-utilization F1 instance.

I'm ambivalent about whether this is how we should be calculating
instance hours (probably good for optimizing cost, less good for
optimizing perf?), but let's at least make it clearer what we're doing
and calculate it correctly!

This commit:
1. Changes the title of the report to `Cost-normalized instance hours`
   which doesn't really clarify what the actual quantity being
   calculated is (it's complicated...), but at least lets you know
   something is being done to the number.
2. Updates all the instance classes. We've changed the instance classes
   of some of our modules, but we didn't update the relative costs. I
   checked each instance class in the `X-module.yaml` files in webapp
   and then updated the cost factor relative to F1 using the linked
   appengine pricing table.
3. Updates the average utilization factors, using the query inline in
   the script to calculate them. For dates, I took one week of data from
   1/18-1/24 (purposfully avoiding the 1/25 outage), and then updated
   the factors in the table. Our average utilization has gone down *a
   lot*, which is maybe concerning in itself, but something to look at
   separately.

My one concern is that this is going to cause a step change in all our
instance hour metrics. I think it's worth doing anyway since these
numbers are just plain wrong right now, but let me know if you think we
shouldn't for some reason. I'll of course PSA it, and I can also push
out the commit on the weekend (it'll take 2 minute, so no big deal) so
that we don't see the step change on a day when we're making other
changes.

Test Plan:
- `gae_dashboard/email_bq_data.py --date 20180125 --dry-run`
- fingers crossed?

Reviewers: amos

Reviewed By: amos

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D41962