Fix another No 2xx on route false positive.
Summary:
We are seeing "No 2xx on route `/_ah/start.*`"
in alerta.

For some reason our logs have the status of
this route as null, though otherwise it appears
that the route is working. I'm not too worried about
that since this is a warmup route called by
appengine when the instance is starting, and perhaps
not everything is running when it's called.

Test Plan: - None

Reviewers: nabil

Subscribers: