Use a status-file to not overload the github servers as much.
We run this script every 10 minutes, and we hit about 15 urls in
quick succession to download all the repo and team information.  We
regularly get at least one 500 error in those downloads.  This change
tries to reduce the amount of fetches we need to make.

In particular, we only need to fetch the list of repositories now (5
fetches).  If that list is the same as last time, then we assume the
teams have already been properly updated and there's nothing to do, so
we early-exit.

Test plan:
I ran, on my local machine:
   ./update_github_teams.py -n
It completed successfully and showed no changes needed to be made.
Then I ran
   ./update_github_teams.py -v -f /var/tmp/ugt.json
and it gave the same (successful) output.  Then I ran
   ./update_github_teams.py -v -f /var/tmp/ugt.json
again, and this time it exited before running:
   Fetching url https://api.github.com/orgs/Khan/teams?per_page=100