Add an email report for log-size.
Summary:
This report says how much we are logging (via logging.info() and
friends), grouped by the first word of the log message.  (Which
usually, but not always, is a good proxy for a single log-message in
our app.)  Now that we pay per byte logged, we want to make sure we're
not accidentally logging a single log message a ton, which is really
easy to do.

Resolves https://khanacademy.atlassian.net/browse/INFRA-1326

Test Plan: Will run manually on toby once this is approved.

Reviewers: amos, benkraft

Reviewed By: benkraft

Subscribers: paulmorgan

Differential Revision: https://phabricator.khanacademy.org/D45150