Add an alert for sentry rate-limiting
Summary:
This is a simple script to notify us if Sentry
starts rate-limiting the prod-js project. We
want to know if this happens because it will make
deploy-time monitoring not work correctly, and we'll
want to adjust the Sentry config to reduce the
error volume.

I will also need to change the toby crontab in the
aws-config repo to call this script once a day. That
will be done in another diff.

Also this script expects to have the Sentry API key
placed in the sentry_api_key file. I'll have to do
that manually on toby.

Test Plan:
- Run the script locally
- Change the script to send an alert regardless of whether there is rate limiting and confirm that
  it sends to sentry.

Reviewers: boris, andrew

Subscribers: