Get info about github repositories and emit it in csv form.
Summary:
This is useful when deciding whether to archive or delete various
repos.  We download info that is helpful for making that decision.
The expectation is the output will be loaded into google spreadsheets
for comments and approval, and then the archiving/deletion decision
will be made based on the output of that process.

Test Plan:
I ran
   ./download_github_info.py
with success, and uploaded the spreadsheet to google docs and all
looks good.

Reviewers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D50431