Script to set sentry issue assignees
Summary:
https://khanacademy.atlassian.net/browse/INFRA-2411

This script finds unassigned sentry issues (in the prod-js
project) and assigns an initiative team based on the URLs
where the issue happens.

The goal is to have all Sentry issues assigned to an initiative.
This will allow us to create reports showing errors by initiative.
It should also help initiatives find relevant issues to work on.

This script uses the same Sentry api key that `check_sentry_rate_limiting.py`
uses.

I will also need to change the toby crontab in the aws-config repo
to call this script once an hour or so. That will be done in another
diff.

Test Plan:
- Ran the script locally and saw that it assigned issues in
  Sentry.

Reviewers: andrew, boris

Reviewed By: boris

Differential Revision: https://phabricator.khanacademy.org/D50441