fetch_amazon_billing has a pip requirement that wasn't in requirements.txt.
The person who set it up must have installed it manually but forgotten
to add it here.  Or maybe it never worked from cron.  :-(  For some
reason it seems to work when run from the commandline, not sure why.

Auditors: benkraft, kai

Test plan:
Will check on the next cron run to make sure there are no error
messages in the log