Simple DoS alert script.
Summary:
https://khanacademy.atlassian.net/browse/INFRA-2781

This script checks for requests that look like simple DoS
attacks and notifies us in slack.

After this is deployed I'll create another diff to add
a crontab entry to run this script every 5 minutes.

Test Plan:
- Run locally. I tested by lowering the threshold so that
we get some results, and then sending a single alert (of
the batch) to #bot-testing

Reviewers: andrew, boris

Differential Revision: https://phabricator.khanacademy.org/D53094