Improve DoS script config
Summary:
Today we've been seeing a number of bogus
DoS alerts. They seem to be due to school districts
accessing Khan from a single IP. Also there
seems to be ocassional browser bugs that
cause a user to hit a Khan URL a bunch of
times in a row.

I'm making the script less sensitive and
I'm adding a link to the stackdriver logs
to make it easier to investigate.

Test Plan:
- Hacked script to post to #bot-testing and
ran it locally to confirm that the logs
link worked.

Reviewers: boris

Subscribers: