Convert scripts to use authoratative who owns data.
Summary:
Instead of using a bunch of manually maintained data
we can now use the "who owns" data generated from webapp
sources nightly.

https://khanacademy.atlassian.net/browse/INFRA-2759

I also changed the check_failing_routes script to
not send to alerta or #infrastructure-alerts but instead
to #infrastructure-sre. Possibly we'll eventually
send these data directly to the teams, but I want
to make sure that it's not too noisy before I turn
that on.

Test Plan:
- Ran `email_bq_data.py --dry-run`
- Ran `check_failing_routes.py --dry-run`
- Ran `set_sentry_assignees.py`
- Changed the sentry script to have it show which teams
  it would assign to own issues that already have ownership
  and spot checked them to make sure that it made sense and
  mostly agreed with existing owners.

Reviewers: boris, andrew

Subscribers: