Add query to check for scratchpad spam.
Summary:
Extension to dos detection script which scans for a large number of new scratchpads created from the same ip.

Scratchpads are found using the bingo_conversion_events and are grouped by ip. Every ip with more than 50 (constant) scratchpads within the past 5 (constant) minutes will be collected and posted to slack. The ip is appended to db-ip.com for geographic lookup, and a sample user's profile is linked using their kaid for manual project checking.

Rate limiting is already done (https://phabricator.khanacademy.org/D53866) but this change will hopefully make it easier examine details about the scratchpads (e.g. geographic location, looking at the projects themselves)

Also did a bit of refactoring to the original dos script so that more constants could be shared with the scratchpad code and the ip sent through the slack message is linked in the same way.

Test Plan:
Ran script locally and checked that messages were posted in slack.
Ran query in BigQuery console and checked that cost was cheap.

{F1431171}

{F1431172}

^ Modified the date information a little since there was no data for the past 5 minutes.

Reviewers: amos, boris

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D54291