Adding error regression alert
Summary:
This is a stacked diff against D54461 which allows us to have regression on routes
spike in error.

We use the same algorithm as performance, but adjust the threshold slightly.

Test Plan:
On 2019-05-03 there was a significant spike in errors on test-prep routes

```
./regresssion_alert.py --date 2019-05-03 --error
```
Should show the alerts

https://khanacademy.slack.com/archives/CCC5YBGA1/p1558043870013100

{F1435187}

https://khanacademy.slack.com/archives/CCC5YBGA1/p1558045581013400

{F1435199}

Example client-error alert (note that this is a manually adjusted threshold - the default threshold is much higher.)

Reviewers: amos, andrew, alexandershah

Reviewed By: amos, andrew

Differential Revision: https://phabricator.khanacademy.org/D54468