Use protobuf streamed logs for DOS alerts.
Summary:
As we are looking to deprecate hourly request logs, we now no longer have a live stream of logs to check for DOS attacks. To address this, we can use the streaming fastly request logs, and run them whenever we check for DOS attacks.

If that query is not modified, it would process around over 100GB, up from the 3GB that this query originally used. To reduce the memory usage, we leverage table decorators (a legacy SQL feature) to scope down the table. This reduces the total query to under 1GB.

Another thing to note is that using the fastly logs doesn't give us the user's KAID, and we can't deduce it from any of its fields. This is used for scratchpad to link us to a sample user's project page.

Test Plan:
- Ran the query in Big Query, inspected the resource usage on the side
- Reduced the limit for DOS / scratchpad alerting, checked that alerts are posted in a slack channel.

Reviewers: amos

Reviewed By: amos

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D56328