Increase DoS Alert Threshold
Summary:
Instead of increasing the period of time we're checking, this increases the number of requests we're looking for per second.

Looking at the past 2 weeks of alerts, all alerts that we determined we attacks were well over 10 req/s.

For most of the alerts we considered browser bugs and legit users, they were under 10 req/s

This bump is being a bit conservative. We could still go higher, but I believe this should eliminate most of the noise, and keep the ones we care about or should investigate

Test Plan:
Intercept the query and make sure it has the new threshold
Compare the threshold to alerts from the past two weeks

Reviewers: amos

Subscribers: boris