Add a cost column to instance hour report.
Summary:
https://khanacademy.atlassian.net/browse/INFRA-288

This change makes it clearer what routes are costing us per day as well as per 1K reqs.
The math is very simple, but I think that this will be clearer to human readers.

Note: we multiply hours by $ 0.05 since the measures are in normalized instance hours
not actual instance hours.

Test Plan:
- I ran python gae_dashboard/email_bq_data.py --dry-run --report=email_instance_hours
- I also hacked the script to show unencoded data since the above command shows
  the email as base64 so that I could verify that it looks alright.

Reviewers: hushbeck

Reviewed By: hushbeck

Subscribers: boris

Differential Revision: https://phabricator.khanacademy.org/D57539