Group DoS Slack Alerts
Summary:
DoS alerts were spamming the channel and making it hard to have conversations about the attack.

To alleviate this problem, we now group alerts so that it should only spam a single message to the channel every 5 minutes at most. Additionally, if this message is long, Slack will automatically include a "show more" link and clip it to only be a certain height.

To group it more efficiently, we also group the messages by IP. During DDoS attacks, we were seeing a single IP spam with multiple user-agents, which was causing additional messages in the channel, when we mostly care about the IP address, and not the 4 different UAs being used for the attack.

Test Plan: Run script and check results in bot testing slack channel

Reviewers: amos

Subscribers: boris