Fix false positives on DOS query.
Summary:
I've been seeing dos alerts for `/.well-known/apple-app-site-association`
for example:
https://khanacademy.slack.com/archives/C8XGW76FQ/p1581004815255300

This isn't right since this file should be cached.

I looked at our query and found the problem - it expected to have
a cache_status of 'HIT' on the edge node. This isn't always the
case. In my manual testing of fetching the file I saw PASS on the
edge and HIT on the shield.

I changed the query to look for a HIT on any node - grouping by
request id to collect all fastly nodes for a single user request.

Issue: none

Test Plan:
- Ran query manually using
  start time: '2020-02-06T15:55:00'
  end time: '2020-02-06T16:00:00'
  table: [fastly.khanacademy_dot_org_logs_20200206]
- saw that I got no results
- switched hit = 1 and got /.well-known/apple-app-site-association
- Ran query for 23:20-23:35 yesterday and saw that we got a result for `/search` as we expected.

Reviewers: hushbeck, boris

Reviewed By: hushbeck

Differential Revision: https://phabricator.khanacademy.org/D60748