Expand sentry rate limiting to other projects
Summary:
We would like to extend our sentry rate limting to other projects.

With the recent additional traffic, we are getting rate limited for
various projects.  This is going to get noisy, but good to know that
at least it is happening.

Test Plan: - Run locally and observe alerts trigger

Reviewers: amos

Reviewed By: amos

Subscribers: hushbeck

Differential Revision: https://phabricator.khanacademy.org/D62049