Add a script to send daily side-by-side test reports to slack
Summary:
This patch adds a script for piping a daily report of side-by-side
testing errors to slack, to help developers understand what errors are
occurring.

The slack message includes links to a mode dashboard [1] that will
provide a bit more detail for the errors we see for any given
side-by-side error. We're using mode here to allow us to do a bit of
formatting of the unified diff field, but more importantly because
@boris pointed out that there's no way to automatically send devs to the
bigquery cloud console with a prefilled query based on query params.

I also took @boris's advice to send these results to a dedicated channel
rather than automatically sending to team channels. It's very possible
that many members of a team aren't necessarily interested in daily
reports so having a separate channel allows folks to opt in to
notifications and to avoid alert fatigue. I do think it's easier to read
the results if they are grouped by team (and the data was already
available on toby) so I did that here.

[1] https://app.mode.com/editor/khanacademy/reports/19660fc6f110

Jira ticket: https://khanacademy.atlassian.net/browse/INFRA-4775

Test Plan:
After setting up secrets appropriately, I ran the script and saw the
following screenshot.

{F1628561}

After landing this diff, I'll install this on toby and setup a cron
task, then ensure it runs the next couple of days.

Reviewers: csilvers, benkraft, boris

Subscribers: