Add a script to notify, on slack, when fastly config changes happen.
Summary:
We have the #whats-happening slack channel that reports on config
changes that affect our running site.  It's useful for debugging when
we have site failures.  Every deploy and publish posts to that
channel, for instance.

We'd like changes to the active fastly config to notify there as well
-- at least for the public-facing fastly services.  The trick is that
while deploys and publishes go through scripts that we write, that can
post to slack, there are changes to fastly that -- as of today -- can
only be made via the fastly UI.

Rather than forcing people to remember to manually post to
that asks fastly if the version of the service has changed recently,
and posts to slack if so.  It is meant to be run every minute or so,
via cron.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5339

Test Plan:
Installed this manually on toby (in /tmp) and ran it, and saw it post
messages to bot-testing that looked like this:
```
Fastly service *.khanacademy.org was modified at 2020-09-21T20:15:03Z.
New version: 288.
Description: Deploy version: 200921-1245-98be99c852af.
```
And when I ran it a second time, it didn't add anything new.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: hushbeck, #devops

Differential Revision: https://phabricator.khanacademy.org/D66228