Make sure problems with the `find` don't cause the script to fail.
The cron-driver sends email if this script exits with a non-zero exit
code.  We expect failure all the time with the `rmdir` -- which just
tries to rmdir everything whether it has files in it or not -- and
sometimes with the "clean old jobs" (if no old jobs remain to be
cleaned).

I also change the rmdir to be depth-first to properly handle nested
dirs (though we don't have any currently).

Auditors: benkraft

Test plan:
Fingers crossed, will check my email