Create the html files and upload them to gcs.
This is (almost) the last step we need to get it so folks can look at
these html graphs!  (The last step is pointing to this bucket
somewhere.)

I created this gcs bucket manually, and gave it the normal "KA"
permissions, meaning most/all Khan employees should be able to view
it.  I also had it retain files for 90 days.

Test plan:
Fingers crossed, will try it out!