Be more conservative rebuilding html.
We need a new approach for handling .data files for deploys that might
currently be in session.  I'll do that in a future commit, but I want
to paper over existing problems for now.

Auditors: benkraft

Test plan:
Ran download_jenkins_perf_data.sh on toby, then visited
   https://storage.cloud.google.com/ka-jenkins-perf/2020-10-07.8d742feff668376647a58154c0955521cfa85748.html
and it showed all 5 jenkins builds for this deploy, rather than the 3
it was showing earlier.