If we fail to build the html file correctly, delete it.
Before, I was using the `-o` flag to output, and it wouldn't output if
something failed.  But now I output to stdout and compress in bash,
meaning we end up with a (compressed) empty file if the visualizer
script fails to complete.  Clean things up in that case.

Auditors: benkraft

Test plan:
Fingers crossed