Make check_failing_route.py not trigger on false positives.
Summary:
https://khanacademy.slack.com/archives/C8XGW76FQ/p1621339209000100

Today the check_failing_routes script ran for the first time in almost
a year. Fixing it to run regularly is another task. This diff is to make
it not trigger on false positives.

There are two types of false positives:

* Old clients. For example `/graphql/getUrgentNotifications [POST]` is
  being triggered by old clients. I believe that the route has changed.

* Attackers. For example `/graphql/console [POST]` is being accessed
  by attackers.

I made two adjustments:

1. I increased the number of non-bot requests and ips necessary to
   trigger an alert.

2. I added some routes to our expected to be failing list.

We may need aditional tuning in the coming days, but this is good
enough to silence the false positives for now.

Issue: None

Test Plan:
```$ python gae_dashboard/check_failing_routes.py --dry-run --date 20210517
No routes with no 2xx requests for 05/17/21
```

Reviewers: hushbeck, boris

Reviewed By: boris

Differential Revision: https://phabricator.khanacademy.org/D71976