Make WAF alert less noisy. (#21)
## Summary:
Originally when we set up with WAF blocking alert we were worried
about the WAF blocking legit traffic and wanted to be alerted so that
we could validate the the traffic was actually bad.

At this point we have a fair amount of confidence in the WAF blocking
bad traffic and so the alert is currently not useful.

This change makes the alert much less sensitive. Now it will only notify
us if significant traffic is being blocked by the WAF. In these cases
we might want to take additional action such as talk with security folks
or implement an IP block.

I set the threshold to 1% which is a guess of where we might want
it. I looked at a bunch of historical alerts and they were are in the
neighbourhood of 0.1% so this threshold should reduce noise. In terms
of whether 1% is the right amount at which we care, I'm not sure. This
is probably worth discussing in code review.

Issue: https://khanacademy.atlassian.net/browse/INFRA-6953

## Test plan:
- None

Author: latteier

Reviewers: LoveAndCoding, dat-boris

Required Reviewers:

Approved By: LoveAndCoding

Checks:

Pull Request URL: https://github.com/Khan/internal-webserver/pull/21