Don't trigger the DoS alert for backfills.
We've seen this alert trigger for backfills before, and it's happening
again today.
https://khanacademy.slack.com/archives/C8XGW76FQ/p1639597513463100

This change tries to keep the alert from triggering on backfills.  I
use a naive approach which is looking for graphql queries whose name
begins with 'backfill'. Hopefully this catches most of them.

Issue: none

Test plan:
- python dos_alert.py to make sure that there are no syntax errors.